move: ddd 패키지 구조에 맞게, 결제 엔티티, repo 이동
