+++ src/main/java/com/bid/auction/domain/payment/repository/PaymentRepository.java
@@ -3,9 +3,11 @@ package com.bid.auction.domain.payment.repository;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import com.bid.auction.domain.payment.entity.Payment;
 
+@Repository
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
 	Optional<Payment> findByMerchantPaymentUid(String uid);
 }
